requirements.txt
azure-identity
azure-mgmt-resource
azure-mgmt-storage
azure-storage-blob
requests
python-dotenv